- Добавил тесты для ServerController.java - Добавил тесты для ClientService.java - Немного причесал код

+++ common/src/main/java/ru/tochitopor/atm/common/GetBalanceReqstDTO.java
@@ -4,7 +4,7 @@ import lombok.Value;
 
 @Value
 public class GetBalanceReqstDTO {
-    private final int clientId;
-    private final int scoreId;
+    private final long clientId;
+    private final long scoreId;
     private final int pin; // не знаю почему, но когда был PIN большими буквами json не хотел его читать
 }
